Split game logic into its own file
Also, minor fixes.
